requirements.txt
numpy~=1.19.5
opencv-python~=4.5.5.62
Pillow~=8.4.0
matplotlib~=3.3.4